requirements.txt
# EDEN OS — Dependencies
# Version: 1.0 | Phase: ONE

# ═══════════════════════════════════════
# Core Framework
# ═══════════════════════════════════════
fastapi>=0.115.0
uvicorn>=0.30.0
websockets>=12.0
pydantic>=2.9.0

# ═══════════════════════════════════════
# ML / Inference
# ═══════════════════════════════════════
torch>=2.4.0
torchaudio>=2.4.0
torchvision>=0.19.0
transformers>=4.45.0
diffusers>=0.31.0
accelerate>=0.34.0
safetensors>=0.4.5
huggingface-hub>=0.25.0

# ═══════════════════════════════════════
# LivePortrait / Face Processing
# ═══════════════════════════════════════
insightface>=0.7.3
onnxruntime-gpu>=1.19.0
mediapipe>=0.10.14

# ═══════════════════════════════════════
# Voice — ASR + TTS + VAD
# ═══════════════════════════════════════
openai-whisper>=20231117

# ═══════════════════════════════════════
# WebRTC
# ═══════════════════════════════════════
aiortc>=1.9.0

# ═══════════════════════════════════════
# Image / Video Processing
# ═══════════════════════════════════════
opencv-python-headless>=4.10.0
Pillow>=10.4.0
numpy>=1.26.4
scipy>=1.14.0
scikit-image>=0.24.0

# ═══════════════════════════════════════
# Utilities
# ═══════════════════════════════════════
pyyaml>=6.0.2
anthropic>=0.34.0
loguru>=0.7.2

# ═══════════════════════════════════════
# Scholar / Knowledge Engine (Agent 7)
# ═══════════════════════════════════════
yt-dlp>=2024.10.22
chromadb>=0.5.5
sentence-transformers>=3.1.0
trafilatura>=1.12.0
pymupdf>=1.24.10
